feat: rating option
This allows rating or score tables to be used to alter the score

Example: a table with scores between 1 and unknown can be used. The score will be recalculated to a scale of 100. This is optional. But when used it does make it possible to increase results if your goal is to have higher scored items first